Abstract utility functions into a separate file
 - This makes it much easier to get around the main command file which
   was quickly becoming too littered.
 - It's possible that the remind utils will be useful to other commands
   as well in the future.
